Migration from MariaDb to Postgre + Added CustomCurve & integration to RankedScore recalculation with ScoreUtils

+++ src/GuildSaber.Api/Features/Guilds/Categories/CategoryResponses.cs
@@ -3,8 +3,8 @@ namespace GuildSaber.Api.Features.Guilds.Categories;
 public static class CategoryResponses
 {
     public readonly record struct Category(
-        ulong Id,
-        ulong GuildId,
+        int Id,
+        int GuildId,
         CategoryInfo Info
     );
